:recycle: refactor: pastas padronizadas nome minusculos
