+++ components/atoms/infoBox/InfoBoxRoot/index.tsx
