+++ components/atoms/infoBox/InfoBoxRoot/styles.ts
