+++ components/atoms/infoBox/InfoBoxText/index.tsx
