+++ components/atoms/infoBox/InfoBoxText/styles.ts
