+++ components/atoms/infoBox/InfoBoxTitle/index.tsx
