+++ components/atoms/infoBox/index.ts
