+++ components/coumpoud/index.ts
@@ -1,2 +1,2 @@
-export { Header } from "./Header";
-export { Footer } from "./Footer";
+export { Header } from "./header";
+export { Footer } from "./footer";
