+++ validation/scheduling/schedulingSchema.ts
